requirements.txt
authorizenet==1.1.6
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
Flask==3.1.0
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==4.9.4
MarkupSafe==3.0.2
python-dotenv==1.1.0
PyXB-X==1.2.6.3
requests==2.32.3
urllib3==2.4.0
Werkzeug==3.1.3
zipp==3.21.0
